pyproject.toml
[build-system]
requires = ["setuptools", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "dejacode"
version = "5.6.0"
description = "Automate open source license compliance and ensure supply chain integrity"
readme = "README.rst"
requires-python = ">=3.14,<3.15"
license = "AGPL-3.0-only"
license-files = ["LICENSE", "NOTICE"]
authors = [
    { name = "nexB Inc.", email = "info@aboutcode.org" }
]
keywords = [
    "open source", "scan", "license", "package", "dependency",
    "copyright", "filetype", "author", "extract", "licensing",
    "scancode", "scanpipe", "docker", "rootfs", "vm",
    "virtual machine", "pipeline", "code analysis", "container"
]
classifiers = [
    "Development Status :: 5 - Production/Stable",
    "Intended Audience :: Developers",
    "Intended Audience :: Information Technology",
    "Intended Audience :: Legal Industry",
    "Programming Language :: Python",
    "Programming Language :: Python :: 3 :: Only",
    "Programming Language :: Python :: 3.14",
    "Topic :: Utilities"
]
dependencies = [
    # Base configuration tools
    "setuptools==82.0.0",
    "wheel==0.46.3",
    "packaging==26.0",
    "pip==26.0.1",
    # Django
    "Django==6.0.2",
    "asgiref==3.11.1",
    "typing_extensions==4.15.0",
    "sqlparse==0.5.5",
    # Django apps
    "django-crispy-forms==2.5",
    "crispy_bootstrap5==2025.6",
    "django-grappelli==4.0.3",
    "django-filter==25.2",
    "django-registration==3.4",
    "confusable_homoglyphs==3.3.1",
    "django-guardian==3.2.0",
    "django-environ==0.12.0",
    "django-debug-toolbar==6.2.0",
    # CAPTCHA
    "altcha==1.0.0",
    "django_altcha==0.9.0",
    # REST API
    "djangorestframework==3.16.1",
    # API documentation
    "drf-yasg==1.21.14",
    "uritemplate==4.2.0",
    "inflection==0.5.1",
    "pytz==2025.2",
    # Track failed login attempts
    "django-axes==8.3.1",
    # Multi-factor authentication
    "django-otp==1.7.0",
    "qrcode==8.2",
    "pypng==0.20220715.0",
    # Database
    "psycopg==3.3.2",
    # Cache
    "redis==7.1.1",
    # Antivirus
    "clamd==1.0.2",
    # Testing
    "model_bakery==1.23.2",
    # Task queue
    "rq==2.6.1",
    "croniter==6.0.0",
    "django-rq==3.2.2",
    "fakeredis==2.33.0",
    # Libs
    "certifi==2026.1.4",
    "urllib3==2.6.3",
    "python-dateutil==2.9.0.post0",
    "python-mimeparse==2.0.0",
    "PyJWT==2.11.0",
    "natsort==8.4.0",
    "six==1.17.0",
    "requests==2.32.5",
    "idna==3.11",
    "charset-normalizer==3.4.4",
    "PyYAML==6.0.3",
    "cython==3.2.4",
    "zipp==3.23.0",
    "XlsxWriter==3.2.9",
    # Markdown
    "markdown==3.10.2",
    "bleach==6.3.0",
    "bleach_allowlist==1.0.3",
    "webencodings==0.5.1",
    # Authentication
    "oauthlib==3.3.1",
    "python3-openid==3.2.0",
    "requests-oauthlib==2.0.0",
    "defusedxml==0.7.1",
    # LDAP Auth
    "python_ldap==3.4.5",
    "setuptools-scm==9.2.2",
    "pyasn1==0.6.2",
    "pyasn1-modules==0.4.2",
    "django-auth-ldap==5.3.0",
    # license expressions
    "boolean.py==5.0",
    "license-expression==30.4.4",
    # Webhooks
    "django-rest-hooks==1.6.1",
    # django-notifications
    "django_notifications_patched==2.0.0",
    "jsonfield==3.2.0",
    "swapper==1.4.0",
    # AboutCode Toolkit
    "aboutcode_toolkit==11.1.1",
    "click==8.3.1",
    "Jinja2==3.1.6",
    "MarkupSafe==3.0.3",
    "saneyaml==0.6.1",
    "openpyxl==3.1.5",
    "et-xmlfile==2.0.0",
    # PackageURL
    "packageurl-python==0.17.6",
    # Gunicorn
    "gunicorn==25.0.3",
    # SPDX validation
    "jsonschema==4.26.0",
    "jsonschema-specifications==2025.9.1",
    "referencing==0.37.0",
    "rpds-py==0.30.0",
    "attrs==25.4.0",
    # CycloneDX
    "cyclonedx-python-lib==11.6.0",
    "sortedcontainers==2.4.0",
    "py-serializable==2.1.0",
    # Git
    "GitPython==3.1.46",
    "gitdb==4.0.12",
    "smmap==5.0.2",
    # CSAF
    "pydantic==2.12.5",
    "pydantic-core==2.41.5",
    "typing-inspection==0.4.2",
    "maturin==1.11.5",
    "setuptools-rust==1.12.0",
    "annotated-types==0.7.0",
    "semantic-version==2.10.0",
    # OpenVEX
    "msgspec==0.20.0",
    # OpenDocument Format
    "odfdo==3.20.2",
    "lxml==6.0.2",
]

[project.optional-dependencies]
dev = [
    # Linter and Validation
    "ruff==0.15.0",
    # Parallel testing
    "tblib==3.2.2"
]
docs = [
    "Sphinx",
    "furo",
    "doc8",
]

[project.urls]
Homepage = "https://github.com/aboutcode-org/dejacode"
Documentation = "https://dejacode.readthedocs.io/"
Repository = "https://github.com/aboutcode-org/dejacode.git"
Issues = "https://github.com/aboutcode-org/dejacode/issues"
Changelog = "https://github.com/aboutcode-org/dejacode/blob/main/CHANGELOG.rst"

[project.scripts]
dejacode = "dejacode:command_line"

[tool.setuptools.packages.find]
where = ["."]

[tool.ruff]
line-length = 100
target-version = "py313"
exclude = [
    "migrations",
    "bin",
    "data",
    "Script",
    "Lib",
    "lib",
    "lib64",
    "local",
    "var",
    "thirdparty",
    "dejacode_toolkit/openvex/",
]

[tool.ruff.lint]
# Rules: https://docs.astral.sh/ruff/rules/
select = [
    "E",  # pycodestyle
    "W",  # pycodestyle warnings
    "D",  # pydocstyle
    "F",  # Pyflakes
    "UP", # pyupgrade
    "S",  # flake8-bandit
    "I",  # isort
    "C9", # McCabe complexity
]
ignore = ["UP032", "D1", "D203", "D205", "D212", "D400", "D415", "S308"]

[tool.ruff.lint.isort]
force-single-line = true
sections = { django = ["django"] }
section-order = [
    "future",
    "standard-library",
    "django",
    "third-party",
    "first-party",
    "local-folder",
]

[tool.ruff.lint.mccabe]
max-complexity = 16

[tool.ruff.lint.per-file-ignores]
# Do not run bandit on test files.
"**/tests/*" = ["S"]
"dejacode_toolkit/csaf/*" = ["D", "UP", "E501", "F401"]
"dejacode_toolkit/spdx.py" = ["UP"]
"component_catalog/models.py" = ["C901"]
